+++ jest.config.ts
@@ -152,7 +152,7 @@ const config: Config = {
   // snapshotSerializers: [],
 
   // The test environment that will be used for testing
-  testEnvironment: "jsdom",
+  testEnvironment: "./lib/FixJSDOMEnvironment.ts",
 
   // Options that will be passed to the testEnvironment
   // testEnvironmentOptions: {},